+++ src/p2p.rs
@@ -1,9 +1,6 @@
 use crate::consensus::ConsensusCommand;
-use crate::rest::endpoints;
 use crate::utils::conf::Conf;
-use anyhow::{Context, Error, Result};
-use axum::routing::get;
-use axum::Router;
+use anyhow::{Error, Result};
 use tokio::net::TcpListener;
 use tokio::sync::mpsc;
 use tracing::info;
@@ -49,19 +46,4 @@ pub async fn p2p_server(
         peer.handshake().await?;
         peer.start().await
     }
-}
-
-pub async fn rest_server(addr: &str) -> Result<()> {
-    info!("rest listening on {}", addr);
-    let app = Router::new()
-        .route("/getTransaction", get(endpoints::get_transaction))
-        .route("/getBlock", get(endpoints::get_block));
-
-    let listener = tokio::net::TcpListener::bind(&addr)
-        .await
-        .context("Starting rest server")?;
-
-    axum::serve(listener, app)
-        .await
-        .context("Starting rest server")
 }
